requirements.txt
jupyterlab
notebook
bigquery-jupyter-plugin
pyspark
google-cloud
numpy
seaborn
pandas
scikit-learn
google-auth
google-cloud-storage
imblearn
streamlit
google-genai
